feat(api): add globalPlace/division/divisionPlace in solves

+++ packages/types/src/responses/bad-avatar-file-size.ts
@@ -5,6 +5,6 @@ export const BadAvatarFileSize = response('badAvatarFileSize', {
   status: 400,
   message: 'The avatar file is too large.',
   data: z.object({
-    maxSize: z.number(),
+    maxSize: z.number().int(),
   }),
 })
