+++ packages/types/src/responses/bad-rate-limit.ts
@@ -5,6 +5,6 @@ export const BadRateLimit = response('badRateLimit', {
   status: 429,
   message: 'You are trying this too fast.',
   data: z.object({
-    timeLeft: z.number(),
+    timeLeft: z.number().int(),
   }),
 })
